# apps/sps-validator-ui/openapi.yaml
openapi: 3.1.0
info:
    title: Splinterlands Validator
    description: "This is a Splinterlands Validator API Server"
    contact:
        email: thespsdao@gmail.com
    license:
        name: MIT
        url: https://choosealicense.com/licenses/mit/
    version: 1.0.0
servers:
    - url: http://localhost:3333/
paths:
...
    /account/{account}:
        get:
            tags:
                - Default
            summary: Get the account
            operationId: getAccount
            parameters:
                - in: path
                  name: account
                  schema:
                      type: string
                  required: true
                  description: Account name
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Account"
    /account/{account}/transactions:
        get:
            tags:
                - Default
            summary: Get the transactions of the account
            operationId: getAccountTransactions
            parameters:
                - in: path
                  name: account
                  schema:
                      type: string
                  required: true
                  description: Account name
                - in: query
                  name: limit
                  schema:
                      type: number
                  description: The number of results to return
                - in: query
                  name: cursor_block_num
                  schema:
                      type: number
                  description: The last block number of the previous page
                - in: query
                  name: cursor_index
                  schema:
                      type: number
                  description: The last transaction index in the block_num of the previous page
                - in: query
                  name: sort
                  schema:
                      type: string
                      enum: ["asc", "desc"]
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Transaction"
    /balances:
        get:
            tags:
                - Default
            summary: Gets the balances of account specified in the query params
            description: Returns the list of balances for the specified account name. Will return empty array if the account name was not found.
            operationId: getBalances
            parameters:
                - in: query
                  name: account
                  schema:
                      type: string
                  required: true
                  description: Account name
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Balances"
    /tokens/{token}:
        get:
            tags:
                - Default
            summary: Gets balances of the token
            description: Returns the list of accounts with a balance of the specified token.
            operationId: getBalancesByToken
            parameters:
                - in: path
                  name: token
                  schema:
                      type: string
                  required: true
                  description: Name of the token
                - in: query
                  name: limit
                  schema:
                      type: number
                  description: The number of results to return
                - in: query
                  name: skip
                  schema:
                      type: number
                  description: The number of results to skip
                - in: query
                  name: systemAccounts
                  schema:
                      type: boolean
                  description: Include system accounts?
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BalancesCount"
    /tokens:
        get:
            tags:
                - Default
            summary: Gets balances of the token
            description: Returns the list of accounts with a balance of the specified token.
            operationId: getBalancesByTokens
            parameters:
                - in: query
                  name: token
                  schema:
                      type: string
                  required: true
                  description: Name of the tokens. Can be specified multiple times to get balances for multiple tokens.
                - in: query
                  name: limit
                  schema:
                      type: number
                  description: The number of results to return
                - in: query
                  name: skip
                  schema:
                      type: number
                  description: The number of results to skip
                - in: query
                  name: systemAccounts
                  schema:
                      type: boolean
                  description: Include system accounts?
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BalancesCount"
    /validators:
        get:
            tags:
                - Default
            summary: Gets the list of validators
            operationId: getValidators
            parameters:
                - in: query
                  name: limit
                  schema:
                      type: integer
                - in: query
                  name: skip
                  schema:
                      type: integer
                - in: query
                  name: search
                  schema:
                      type: string
                - in: query
                  name: active
                  schema:
                      type: boolean
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Validators"
    /validator:
        get:
            tags:
                - Default
            summary: Gets a validator
            operationId: getValidator
            parameters:
                - in: query
                  name: account
                  schema:
                      type: string
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Validator"
    /votes_by_account:
        get:
            tags:
                - Default
            summary: Gets the votes performed by the specified account
            operationId: getVotesByAccount
            parameters:
                - in: query
                  name: account
                  schema:
                      type: string
                  required: true
                  description: Account name
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidatorVotes"
    /votes_by_validator:
        get:
            tags:
                - Default
            summary: Gets the votes given to the specified validator
            operationId: getVotesByValidator
            parameters:
                - in: query
                  name: validator
                  schema:
                      type: string
                  required: true
                  description: Validator
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidatorVotes"
...
    /pool/{poolName}:
        get:
            tags:
                - Default
            summary: Gets the pool parameters of a specific reward pool
            operationId: getPoolParams
            parameters:
                - in: path
                  name: poolName
                  schema:
                      type: string
                  required: true
                  description: Name of the reward pool
            responses:
                200:
                    description: Correctly configured pool params
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PoolSettings"
                404:
                    description: Pool not found
                503:
                    description: Pool found, but currently not configured correctly
    /pool/{poolName}/reward_debt:
        get:
            tags:
                - Default
            summary: Gets the reward debt for an account in a specific reward pool
            operationId: getRewardDebt
            parameters:
                - in: path
                  name: poolName
                  schema:
                      type: string
                  required: true
                  description: Name of the reward pool
                - in: query
                  name: account
                  schema:
                      type: string
                  required: true
                  description: Account name
            responses:
                200:
                    description: Current reward debt of account in pool
                    content:
                        application/json:
                            schema:
                                type: number
                204:
                    description: No reward debt for this account in pool
                400:
                    description: Missing account query parameter
                404:
                    description: Pool not found
    /transactions/{blockNum}/token_transfer:
        get:
            tags:
                - Default
            summary: Gets the token transfers performed in the block
            operationId: getTokenTransferTransactions
            parameters:
                - in: path
                  name: blockNum
                  schema:
                      type: integer
                  required: true
                  description: Block number
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TokenTransferTransactions"
                400:
                    description: User error
    /transaction:
        get:
            tags:
                - Default
            summary: Looks up a transaction by ID
            operationId: getTransaction
            parameters:
                - in: query
                  name: id
                  schema:
                      type: string
                  required: true
                  description: Transaction ID
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Transaction"
...
components:
    schemas:
        Account:
            type: object
            required:
                - name
            properties:
                name:
                    type: string
                authority:
                    type: object
...
        Balances:
            type: array
            items:
                $ref: "#/components/schemas/Balance"
        BalancesCount:
            type: object
            properties:
                count:
                    type: integer
                    format: int64
                balances:
                    type: array
                    items:
                        $ref: "#/components/schemas/Balance"
...
        Transaction:
            type: object
            required:
                - id
                - block_id
                - prev_block_id
                - block_num
                - type
                - player
                - index
            properties:
                id:
                    type: string
                block_id:
                    type: string
                prev_block_id:
                    type: string
                block_num:
                    type: integer
                type:
                    type: string
                player:
                    type: string
                data:
                    type: string
                success:
                    type: boolean
                error:
                    type: string
                created_date:
                    type: string
                    format: date-time
                result:
                    type: string
                index:
                    type: integer
...